Restaurant Point of Sale Application
As a group, several classmates and I designed a point of sale
application for restaurants. It has two components: an interface for
the diners and an interface for the restaurant. Instead of exchanging
cards with a waiter, diners can split up their checks how they would
like through, add a tip and more, with a few taps of their smartphones.